+++ app/pdfs/pdf/invoice.rb
@@ -6,7 +6,8 @@ class PDF::Invoice < Prawn::Document
   include ActionView::Helpers::NumberHelper
   include ApplicationHelper
 
-  def initialize(invoice, subscription_expiration_date)
+  # @param invoice [Invoice]
+  def initialize(invoice)
     super(margin: 70)
 
     # fonts
@@ -42,8 +43,8 @@ class PDF::Invoice < Prawn::Document
         text I18n.t('invoices.invoice_reference', REF: invoice.reference), leading: 3
       end
       text I18n.t('invoices.code', CODE: Setting.get('invoice_code-value')), leading: 3 if Setting.get('invoice_code-active')
-      if invoice.main_item.object_type != WalletTransaction.name
-        order_number = invoice.main_item.object_type == OrderItem.name ? invoice.main_item.object.order.reference : invoice.order_number
+      if invoice.main_item&.object_type != WalletTransaction.name
+        order_number = invoice.main_item&.object_type == OrderItem.name ? invoice.main_item&.object&.order&.reference : invoice.order_number
         text I18n.t('invoices.order_number', NUMBER: order_number), leading: 3
       end
       if invoice.is_a?(Avoir)
@@ -80,60 +81,7 @@ class PDF::Invoice < Prawn::Document
       invoice.invoice_items.each do |item|
         price = item.amount.to_i / 100.00
 
-        details = invoice.is_a?(Avoir) ? "#{I18n.t('invoices.cancellation')} - " : ''
-
-        if item.object_type == Subscription.name
-          subscription = item.object
-          if invoice.main_item.object_type == 'OfferDay'
-            details += I18n.t('invoices.subscription_extended_for_free_from_START_to_END',
-                              START: I18n.l(invoice.main_item.object.start_at.to_date),
-                              END: I18n.l(invoice.main_item.object.end_at.to_date))
-          else
-            subscription_end_at = case subscription_expiration_date
-                                  when Time
-                                    subscription_expiration_date
-                                  when String
-                                    DateTime.parse(subscription_expiration_date)
-                                  else
-                                    subscription.expiration_date
-                                  end
-            subscription_start_at = subscription_end_at - subscription.plan.duration
-            details += I18n.t('invoices.subscription_NAME_from_START_to_END',
-                              NAME: item.description,
-                              START: I18n.l(subscription_start_at.to_date),
-                              END: I18n.l(subscription_end_at.to_date))
-          end
-
-        elsif item.object_type == Reservation.name
-          case invoice.main_item.object.try(:reservable_type)
-            ### Machine reservation
-          when 'Machine'
-            details += I18n.t('invoices.machine_reservation_DESCRIPTION', DESCRIPTION: item.description)
-          when 'Space'
-            details += I18n.t('invoices.space_reservation_DESCRIPTION', DESCRIPTION: item.description)
-          ### Training reservation
-          when 'Training'
-            details += I18n.t('invoices.training_reservation_DESCRIPTION', DESCRIPTION: item.description)
-          ### events reservation
-          when 'Event'
-            details += I18n.t('invoices.event_reservation_DESCRIPTION', DESCRIPTION: item.description)
-            # details of the number of tickets
-            if invoice.main_item.object.nb_reserve_places.positive?
-              details += "\n  #{I18n.t('invoices.full_price_ticket', count: invoice.main_item.object.nb_reserve_places)}"
-            end
-            invoice.main_item.object.tickets.each do |t|
-              details += "\n #{I18n.t('invoices.other_rate_ticket',
-                                      count: t.booked,
-                                      NAME: t.event_price_category.price_category.name)}"
-            end
-          else
-            details += item.description
-          end
-        else
-          details += item.description
-        end
-
-        data += [[details, number_to_currency(price)]]
+        data += [[Invoices::ItemLabelService.build(invoice, item), number_to_currency(price)]]
         total_calc += price
         total_ht += item.net_amount
         total_vat += item.vat
@@ -221,71 +169,18 @@ class PDF::Invoice < Prawn::Document
 
       # payment details
       move_down 20
-      if invoice.is_a?(Avoir)
-        payment_verbose = "#{I18n.t('invoices.refund_on_DATE', DATE: I18n.l(invoice.avoir_date.to_date))} "
-        case invoice.payment_method
-        when 'stripe'
-          payment_verbose += I18n.t('invoices.by_card_online_payment')
-        when 'cheque'
-          payment_verbose += I18n.t('invoices.by_cheque')
-        when 'transfer'
-          payment_verbose += I18n.t('invoices.by_transfer')
-        when 'cash'
-          payment_verbose += I18n.t('invoices.by_cash')
-        when 'wallet'
-          payment_verbose += I18n.t('invoices.by_wallet')
-        when 'none'
-          payment_verbose = I18n.t('invoices.no_refund')
-        else
-          Rails.logger.error "specified refunding method (#{payment_verbose}) is unknown"
-        end
-        payment_verbose += " #{I18n.t('invoices.for_an_amount_of_AMOUNT', AMOUNT: number_to_currency(total))}"
-      else
-        # subtract the wallet amount for this invoice from the total
-        if invoice.wallet_amount
-          wallet_amount = invoice.wallet_amount / 100.00
-          total -= wallet_amount
-        else
-          wallet_amount = nil
-        end
-
-        # payment method
-        payment_verbose = if invoice.paid_by_card?
-                            I18n.t('invoices.settlement_by_debit_card')
-                          else
-                            I18n.t('invoices.settlement_done_at_the_reception')
-                          end
-
-        # if the invoice was 100% payed with the wallet ...
-        payment_verbose = I18n.t('invoices.settlement_by_wallet') if total.zero? && wallet_amount
-
-        payment_verbose += " #{I18n.t('invoices.on_DATE_at_TIME',
-                                      DATE: I18n.l(invoice.created_at.to_date),
-                                      TIME: I18n.l(invoice.created_at, format: :hour_minute))}"
-        if total.positive? || !invoice.wallet_amount
-          payment_verbose += " #{I18n.t('invoices.for_an_amount_of_AMOUNT', AMOUNT: number_to_currency(total))}"
-        end
-        if invoice.wallet_amount
-          payment_verbose += if total.positive?
-                               " #{I18n.t('invoices.and')} #{I18n.t('invoices.by_wallet')} " \
-                                 "#{I18n.t('invoices.for_an_amount_of_AMOUNT', AMOUNT: number_to_currency(wallet_amount))}"
-                             else
-                               " #{I18n.t('invoices.for_an_amount_of_AMOUNT', AMOUNT: number_to_currency(wallet_amount))}"
-                             end
-        end
-      end
-      text payment_verbose
+      text Invoices::PaymentDetailsService.build(invoice, total)
 
       # important information
       move_down 40
-      txt = parse_html(Setting.get('invoice_text'))
+      txt = parse_html(Setting.get('invoice_text').to_s)
       txt.each_line do |line|
         text line, style: :bold, inline_format: true
       end
 
       # address and legals information
       move_down 40
-      txt = parse_html(Setting.get('invoice_legals'))
+      txt = parse_html(Setting.get('invoice_legals').to_s)
       txt.each_line do |line|
         text line, align: :right, leading: 4, inline_format: true
       end
@@ -302,21 +197,6 @@ class PDF::Invoice < Prawn::Document
   end
 
   private
-
-  def reservation_dates_verbose(slot)
-    if slot.start_at.to_date == slot.end_at.to_date
-      "- #{I18n.t('invoices.on_DATE_from_START_to_END',
-                  DATE: I18n.l(slot.start_at.to_date),
-                  START: I18n.l(slot.start_at, format: :hour_minute),
-                  END: I18n.l(slot.end_at, format: :hour_minute))}\n"
-    else
-      "- #{I18n.t('invoices.from_STARTDATE_to_ENDDATE_from_STARTTIME_to_ENDTIME',
-                  STARTDATE: I18n.l(slot.start_at.to_date),
-                  ENDDATE: I18n.l(slot.start_at.to_date),
-                  STARTTIME: I18n.l(slot.start_at, format: :hour_minute),
-                  ENDTIME: I18n.l(slot.end_at, format: :hour_minute))}\n"
-    end
-  end
 
   ##
   # Remove every unsupported html tag from the given html text (like <p>, <span>, ...).